+++ product-reservation-service/src/main/java/com/monkey/productreservationservice/domain/enums/ProductReservationStatus.java
@@ -1,4 +1,4 @@
-package com.monkey.productreservationservice.domain.entity;
+package com.monkey.productreservationservice.domain.enums;
 
 public enum ProductReservationStatus {
     PENDING_PICKUP,
